5767 - Separate global and country repository
